Web - Added support for finding mob spawnable blocks.  See the '--check-spawn' option; Web - For features, Pos element is now clickable to go to the indicated layer; Web - Moved output images to subdirectory 'images/'; Web - Enforce layer z-index ordering to improve interaction of the many layers we now have; XML - Added flags for blocks: opaque, liquid, spawnable; XML - Added stair blockvariants; XML - Fixed parsing of boolean values; Added global histograms for biomes and blocks to the log file-- histograms are sorted and show percentages; Web - Updated OpenLayers from v3.11.1 to v3.11.2;

+++ mcpe_viz.nbt.h
@@ -27,7 +27,7 @@ namespace mcpe_viz {
   typedef std::vector< MyNbtTag > MyNbtTagList;
 
 
-  std::string makeGeojsonHeader(int32_t ix, int32_t iy);
+  std::string makeGeojsonHeader(double ix, double iy);
   
   int parseNbt( const char* hdr, const char* buf, int bufLen, MyNbtTagList& tagList );
